working on pots page

+++ src/app/pages/delete/delete/confirm-delete/confirm-delete.component.ts
@@ -1,5 +1,6 @@
 import { Component, EventEmitter, inject, Input, Output } from '@angular/core';
-import { BudgetsService } from '../../../services/budgets/budgets.service';
+import { BudgetsService } from '../../../../services/budgets/budgets.service';
+import { PotsService } from '../../../../services/pots/pots.service';
 
 @Component({
   selector: 'app-confirm-delete',
@@ -9,12 +10,15 @@ import { BudgetsService } from '../../../services/budgets/budgets.service';
   styleUrl: './confirm-delete.component.css',
 })
 export class ConfirmDeleteComponent {
+  @Input() type!: string;
   @Input() index!: number;
   @Output() exitPage = new EventEmitter<Event>();
   budgetService = inject(BudgetsService);
+  potService = inject(PotsService);
 
-  deleteBudget() {
-    this.budgetService.removeBudget(this.index);
+  deleteResource() {
+    if (this.type === 'budgets') this.budgetService.removeBudget(this.index);
+    else if (this.type === 'pots') this.potService.removePot(this.index);
     this.exitPage.emit();
   }
 }
